feat: move delete
