+++ src/console/src/proposals.rs
@@ -1,17 +1,8 @@
 use crate::metadata::update_releases_metadata;
-use crate::storage::state::stable::{
-    delete_asset_stable, delete_content_chunks_stable, get_assets_stable,
-};
-use crate::store::stable::get_proposal;
 use crate::strategies_impls::cdn::{CdnHeap, CdnStable, CdnWorkflow};
 use candid::Principal;
-use junobuild_cdn::proposals::errors::{
-    JUNO_ERROR_PROPOSALS_CANNOT_DELETE_ASSETS,
-    JUNO_ERROR_PROPOSALS_CANNOT_DELETE_ASSETS_INVALID_STATUS, JUNO_ERROR_PROPOSALS_UNKNOWN_TYPE,
-};
 use junobuild_cdn::proposals::{CommitProposal, CommitProposalError};
-use junobuild_cdn::proposals::{Proposal, ProposalId, ProposalStatus, ProposalType};
-use junobuild_shared::utils::principal_not_equal;
+use junobuild_cdn::proposals::{Proposal, ProposalId, ProposalType};
 
 pub fn init_proposal(
     caller: Principal,
@@ -35,18 +26,7 @@ pub fn delete_proposal_assets(
     caller: Principal,
     proposal_ids: &Vec<ProposalId>,
 ) -> Result<(), String> {
-    for proposal_id in proposal_ids {
-        let proposal = get_proposal(proposal_id);
-
-        match proposal {
-            None => {
-                return Err(JUNO_ERROR_PROPOSALS_CANNOT_DELETE_ASSETS.to_string());
-            }
-            Some(proposal) => secure_delete_proposal_assets(caller, proposal_id, &proposal)?,
-        }
-    }
-
-    Ok(())
+    junobuild_cdn::proposals::delete_proposal_assets(&CdnStable, caller, proposal_ids)
 }
 
 pub fn post_commit_assets(proposal: &Proposal) -> Result<(), String> {
@@ -56,49 +36,6 @@ pub fn post_commit_assets(proposal: &Proposal) -> Result<(), String> {
             return update_releases_metadata(options);
         }
     }
-
-    Ok(())
-}
-
-pub fn secure_delete_proposal_assets(
-    caller: Principal,
-    proposal_id: &ProposalId,
-    proposal: &Proposal,
-) -> Result<(), String> {
-    // The one that uploaded the assets can remove those.
-    if principal_not_equal(caller, proposal.owner) {
-        return Err(JUNO_ERROR_PROPOSALS_CANNOT_DELETE_ASSETS.to_string());
-    }
-
-    if proposal.status == ProposalStatus::Open {
-        return Err(format!(
-            "{} ({:?})",
-            JUNO_ERROR_PROPOSALS_CANNOT_DELETE_ASSETS_INVALID_STATUS, proposal.status
-        ));
-    }
-
-    assert_known_proposal_type(proposal)?;
-
-    let assets = get_assets_stable(proposal_id);
-
-    for (key, asset) in assets {
-        for (_, encoding) in asset.encodings.iter() {
-            delete_content_chunks_stable(&encoding.content_chunks);
-        }
-
-        delete_asset_stable(&key);
-    }
-
-    Ok(())
-}
-
-fn assert_known_proposal_type(proposal: &Proposal) -> Result<(), String> {
-    #[allow(unreachable_patterns)]
-    match &proposal.proposal_type {
-        ProposalType::AssetsUpgrade(_) => (),
-        ProposalType::SegmentsDeployment(_) => (),
-        _ => return Err(JUNO_ERROR_PROPOSALS_UNKNOWN_TYPE.to_string()),
-    };
 
     Ok(())
 }
