+++ src/libs/cdn/src/proposals/mod.rs
@@ -7,6 +7,7 @@ pub use state::stable;
 pub use state::types::*;
 pub use types::*;
 pub use workflows::commit::*;
+pub use workflows::delete::*;
 pub use workflows::errors;
 pub use workflows::init::*;
 pub use workflows::pre_commit_assets::*;
