+++ src/libs/cdn/src/proposals/workflows/mod.rs
@@ -1,5 +1,6 @@
 mod assert;
 pub mod commit;
+pub mod delete;
 pub mod errors;
 pub mod init;
 pub mod pre_commit_assets;
